+++ packages/wasm/lib/recipe/page.js
@@ -651,11 +651,9 @@ export function createPageMethods(
       if (this._editingPage) {
         if (this._contextState === "active-edit") {
           this._page.endContext();
-          this._flushAnnotations();
-          this._page.writePage();
-        } else {
-          this._flushAnnotations();
         }
+        this._flushAnnotations();
+        this._page.writePage();
         this._pageContext = null;
         this._page = null;
         this._editingPage = false;
@@ -1049,8 +1047,6 @@ export function createPageMethods(
     pauseContext: function () {
       if (this._contextState === "active-edit") {
         this._page.endContext();
-        this._page.writePage();
-        this._page = null;
         this._pageContext = null;
         this._contextState = "paused-edit";
       } else if (this._contextState === "active-new") {
@@ -1079,10 +1075,6 @@ export function createPageMethods(
      */
     resumeContext: function () {
       if (this._contextState === "paused-edit") {
-        this._page = this.writer.createPageModifier(
-          this._activePageNumber - 1,
-          true,
-        );
         this._pageContext = this._page.startContext().getContext();
         this._resumePageRotation();
         this._contextState = "active-edit";